create tower external helper
